+++ CHANGELOG.md
@@ -11,6 +11,7 @@ dev
  * ons.platform: Only prompt input cursor for supported input types. ([#2803](https://github.com/OnsenUI/OnsenUI/issues/2803)).
  * core: Fix animations not working for platform-specific animations.
  * ons-fab: Fix fab not being positioned relative to tab page. ([#2810](https://github.com/OnsenUI/OnsenUI/issues/2810)).
+ * core: Fix setImmediate not defined for ESM bug.
 
 2.11.1
 ---
