+++ core/src/ons/internal/internal.js
@@ -101,20 +101,6 @@ window.document.addEventListener('_templateloaded', function(e) {
   }
 }, false);
 
-internal.waitDOMContentLoaded(function() {
-  register('script[type="text/ons-template"]');
-  register('script[type="text/template"]');
-  register('script[type="text/ng-template"]');
-  register('template');
-
-  function register(query) {
-    const templates = window.document.querySelectorAll(query);
-    for (let i = 0; i < templates.length; i++) {
-      internal.templateStore.set(templates[i].getAttribute('id'), templates[i].textContent || templates[i].content);
-    }
-  }
-});
-
 /**
  * @param {String} page
  * @return {Promise}
